requirements.txt
Django>=6.0.1,<7.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
drf-spectacular>=0.27.0
django-import-export>=4.0.0
openpyxl>=3.1.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
